Pruebas terminadas y proyecto terminado, solo queda bug de resolver operaciones con decimales

+++ 21911267_ProyectoLenguajes/Evaluator.cpp
@@ -149,7 +149,7 @@ string Evaluator::PostfixToResult(string expre) noexcept(false){
 				d += expression.at(x).at(w);
 				w++;
 			}		
-			operands.push(atof(d.c_str()));
+			operands.push(stof(d.c_str()));
 			d = "";
 			w = 0;
